+++ include/mln/webgpu/context.hpp
@@ -53,7 +53,9 @@ public:
                                               bool persistent = false,
                                               bool ssbo = false) override;
     gfx::UniqueUniformBufferArray createLayerUniformBufferArray() override;
-    gfx::ShaderProgramBasePtr getGenericShader(gfx::ShaderRegistry&, const std::string& name) override;
+    gfx::ShaderProgramBasePtr getGenericShader(gfx::ShaderRegistry&,
+                                               const std::string& name,
+                                               gfx::ProjectionVariant) override;
     TileLayerGroupPtr createTileLayerGroup(int32_t layerIndex, std::size_t initialCapacity, std::string name) override;
     LayerGroupPtr createLayerGroup(int32_t layerIndex, std::size_t initialCapacity, std::string name) override;
     gfx::Texture2DPtr createTexture2D() override;